feat: custom response types

+++ TaskManager/TaskManager.Application/Features/Projects/Handlers/Commands/DeleteProjectCommandHandler.cs
@@ -2,10 +2,11 @@
 using MediatR;
 using TaskManager.Application.Contracts.Persistence;
 using TaskManager.Application.Features.Projects.Requests.Commands;
+using TaskManager.Application.Responses;
 
 namespace TaskManager.Application.Features.Projects.Handlers.Commands
 {
-    public class DeleteProjectCommandHandler : IRequestHandler<DeleteProjectCommand>
+    public class DeleteProjectCommandHandler : IRequestHandler<DeleteProjectCommand, BaseCommandResponse>
     {
         private readonly IProjectRepository _projectRepository;
         private readonly IMapper _mapper;
@@ -16,13 +17,24 @@ namespace TaskManager.Application.Features.Projects.Handlers.Commands
             _mapper = mapper;
         }
 
-        public async Task<Unit> Handle(DeleteProjectCommand request, CancellationToken cancellationToken)
+        public async Task<BaseCommandResponse> Handle(DeleteProjectCommand request, CancellationToken cancellationToken)
         {
+            var response = new BaseCommandResponse();
+
             var project = await _projectRepository.GetAsync(request.Id);
+
+            if(project is null)
+            {
+                response.Success = false;
+                response.Message = $"Project deletion failed. Missing Project with ID {request.Id}";
+                return response;
+            }
 
             await _projectRepository.DeleteAsync(project);
 
-            return Unit.Value;
+            response.Message = "Project deletion successful.";
+
+            return response;
         }
     }
 }
