+++ TaskManager/TaskManager.Application/Features/Todos/Handlers/Commands/DeleteTodoCommandHandler.cs
@@ -2,10 +2,11 @@
 using MediatR;
 using TaskManager.Application.Contracts.Persistence;
 using TaskManager.Application.Features.Todos.Requests.Commands;
+using TaskManager.Application.Responses;
 
 namespace TaskManager.Application.Features.Todos.Handlers.Commands
 {
-    public class DeleteTodoCommandHandler : IRequestHandler<DeleteTodoCommand>
+    public class DeleteTodoCommandHandler : IRequestHandler<DeleteTodoCommand, BaseCommandResponse>
     {
         private readonly ITodoRepository _todoRepository;
         private readonly IMapper _mapper;
@@ -16,13 +17,24 @@ namespace TaskManager.Application.Features.Todos.Handlers.Commands
             _mapper = mapper;
         }
 
-        public async Task<Unit> Handle(DeleteTodoCommand request, CancellationToken cancellationToken)
+        public async Task<BaseCommandResponse> Handle(DeleteTodoCommand request, CancellationToken cancellationToken)
         {
+            var response = new BaseCommandResponse();
+
             var todo = await _todoRepository.GetAsync(request.Id);
+
+            if(todo is null)
+            {
+                response.Success = false;
+                response.Message = $"Todo deletion failed. Missing Todo with ID {request.Id}";
+                return response;
+            }
 
             await _todoRepository.DeleteAsync(todo);
 
-            return Unit.Value;
+            response.Message = "Todo deletion successful.";
+
+            return response;
         }
     }
 }
